formatted solution code and changed class name
